+++ Plugins/WorldAssembly/Source/NexusWorldAssembly/Private/Cell/NCellActor.cpp
@@ -90,7 +90,7 @@ void ANCellActor::InitializeFromProxy(ANCellLevelInstance* LevelInstance)
 	OnInitializedFromProxy.ExecuteIfBound(LevelInstance);
 
 	// Level-based callback
-	for (const auto& Actor : InitializeCallbackActors)
+	for (const auto& Actor : CellInitializedTargets)
 	{
 		INCellInitialized::Execute_OnInitializedFromProxy(Actor, LevelInstance);
 	}